made sweeping changes to the infrastructure of the application. Resource dictionaries for farm are now in <string, int> format instead of <IPlant, int> or <IAnimal, int>. Still working on fixing the processing functions. I'm also trying to make some utility functions for identifying kvps in dictionaries.
