+++ src/Models/ProductInventory.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,6 +9,8 @@ namespace Trestlebridge.Models
 {
     public class ProductInventory
     {
+        internal IEnumerable<PropertyInfo> p;
+
         // Fields
         private double _meat;
         private double _eggs;
